requirements.txt
# Piggy Voice Research Agent

python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.28.0
SpeechRecognition>=3.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
